pyproject.toml
[project]
name = "midnight-oil"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "authlib>=1.6.5",
    "duckdb>=1.4.2",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
]
